NUN-6431: Homepage Empty State Dark mode

+++ nunchuk-main/src/main/java/com/nunchuk/android/main/components/tabs/wallet/emptystate/EmptyStateProvider.kt
@@ -16,7 +16,7 @@ data class WizardData(
     val buttonText: String,
     val buttonAction: () -> Unit,
     val imageResId: Int,
-    val backgroundColor: Long
+    val backgroundColor: Int
 )
 
 data class KeyWalletEntryData(
